+++ R/calculate_geodesic_distances.R
@@ -1,10 +1,10 @@
 #' Calculate geodesic distances between cells in a trajectory
 #'
-#' Will calculate geodesic distances along a trajectory. To speed things up, only the distances with a set of waypoint cells are calculated.
+#' Will calculate geodesic distances between cells within a trajectory. To speed things up, only the distances with a set of waypoint cells are calculated.
 #'
-#' The geodesic distance takes into acount regions of delayed commitment.
+#' The geodesic distance takes into account the length of an edge regions of delayed commitment.
 #'
-#' @param trajectory The trajectory
+#' @inheritParams common_param
 #' @param waypoint_cells A vector of waypoint cells. Only the geodesic distances between waypoint cells and all other cells will be calculated.
 #' @param waypoint_milestone_percentages The milestone percentages of non-cell waypoints, containing waypoint_id, milestone_id and percentage columns
 #'
@@ -12,6 +12,10 @@
 #'
 #' @importFrom igraph graph_from_data_frame neighborhood E distances
 #' @importFrom reshape2 acast melt
+#'
+#' @examples
+#' data(example_dataset)
+#' geodesic_distances <- calculate_geodesic_distances(example_dataset)
 #'
 #' @export
 calculate_geodesic_distances <- function(
@@ -183,6 +187,6 @@ compute_tented_geodesic_distances <- function(
   waypoint_cells = NULL,
   waypoint_milestone_percentages = NULL
 ) {
-  warning("compute_tented_geodesic_distances in deprecated and will be removed in future releases. Use calculate_geodesic_distances instead.")
+  warning("compute_tented_geodesic_distances is deprecated and will be removed in future releases. Use calculate_geodesic_distances instead.")
   calculate_geodesic_distances(trajectory, waypoint_cells, waypoint_milestone_percentages)
 }