+++ R/method_create_ti_method_r.R
@@ -78,9 +78,9 @@ create_ti_method_r <- function(
   # only give args that are requested by the function
   args <- args[intersect(names(args), names(formals(method$run$run_fun)))]
 
-  model <- do.call(method$run$run_fun, args)
+  trajectory <- do.call(method$run$run_fun, args)
 
-  model
+  trajectory
 }