+++ R/wrap_gather_cells_at_milestones.R
@@ -1,6 +1,6 @@
 #' "Gather" cells to their closest milestones
 #'
-#' @param trajectory The trajectory
+#' @inheritParams common_param
 #'
 #' @keywords adapt_trajectory
 #'